backend.old/requirements.txt
Flask==3.1.0
Flask-Caching==2.3.0
Flask-Cors==5.0.0
Flask-Limiter==3.9.2
Flask-SQLAlchemy==3.1.1
flask-talisman==1.1.0
flasgger==0.9.7.1
gevent==24.2.1
loguru==0.7.3
psycopg==3.2.3
psycopg-binary==3.2.3
psycopg-pool==3.2.4
PyJWT==2.10.1
python-dotenv==1.0.1
python-telegram-bot==21.9
redis==5.2.1
requests==2.32.3
tonsdk==1.0.15
prometheus-flask-exporter==0.23.0
prometheus-client==0.19.0
fastapi==0.115.6
uvicorn[standard]==0.34.0
asyncpg==0.29.0
SQLAlchemy[asyncio]==2.0.27
PyNaCl==1.5.0
